Fix subtle crash in sysreqD() because of modified EBX
I forgot that EBX is used inside beginJitCode() and
that it must be re-initialized because the caller i.e.
sysreqD() can modify it. Luckily (or rather not so) it
worked fine on Windows (MSVC) with any game mode so I
was almost certain it must go well on Linux as well...