Shorten STACK/HEAP instructions when possible
This is achieved by finding out whether one of the
subsequent instructions needs ALT before it is over-
written by another instruction. If yes, we have to
do the full job i.e. assign STK to ALT, and if no
we can omit that part and make the resulting JIT code
2 instructions shorter.