Emit direct native calls in non-debug mode
Call native functions directly by address instead of invoking them indirectly
via a table index. I couldn't see any reason for not doing so unless you want
to intercept native calls for some reason. But that is not possible to do from
within a script itself in the current version because all scripts are compiled
only once and before they are executed, and I'm not aware of any plugins that
would want to do something like this (except maybe profiler), so it's sort of
useless...

It could be useful for debugging, though.