Make it possible to prepare code for JIT compilation (issue #35)
It's well known that it doesn't make much sense to run a self-modifying
code under the JIT plugin because changes are not recompiled. But with this new OnJITCompile() callback the code can run all its dirty stuff  in there (including rewriting itself).

Note: Return value of 0 in OnJITCompile() effectively disables the on-going compilation. I don't know where this can be useful though.