Don't clear DF before using rep XXX
There's no need to do so as it's guaranteed to be unset upon function entry.

http://www.agner.org/optimize/calling_conventions.pdf

"The rules for the direction flag is the same in all systems. The direction
flag is cleared by default. If the direction flag is set, then it must be
cleared again before any call or return. Some compilers and subroutine
libraries rely on the direction flag always being clear (Microsoft, Watcom,
Digital Mars) while other systems use the double-safe strategy of always
leaving the direction flag cleared, but not relying on receiving it cleared
(Borland, Gnu)."